requirements.txt
joblib==1.0.1
numpy==1.20.3
pandas==1.2.4
Pillow==8.2.0
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
scikit-learn==0.24.2
scipy==1.7.0
six==1.16.0
threadpoolctl==2.1.0
torch==1.9.0+cu111
torchaudio==0.9.0
torchvision==0.10.0+cu111
tqdm==4.61.1
typing-extensions==3.10.0.0
